requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-PyMongo==2.3.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.2.1
WTForms==3.1.0
pymongo==4.5.0
Werkzeug==2.3.7
python-dotenv==1.0.0
Pillow==10.0.0
requests==2.31.0
bcrypt==4.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
gunicorn==21.2.0
email-validator==2.0.0
python-magic==0.4.27
pytz==2023.3
dnspython==2.4.2
blinker==1.6.2
